Filter ApplicationFrameWindows that don't actively host programs #54
The ApplicationFrameWindows that host Windows Store Apps like to
hang around in Windows 10 even after the underlying program has been
closed. A way to figure out if the ApplicationFrameWindow is
currently hosting an application is to check if it has a property called
"ApplicationViewCloakType", and that the value != 1.

I've stumbled upon these values of "ApplicationViewCloakType":
 0 = Program is running on current virtual desktop
 1 = Program is not running
 2 = Program is running on a different virtual desktop