Bugfix: auto-switch triggered on quick release of custom shortcut
The above else clause causes Switcheroo to auto-switch on a too quickly released custom shortcut (i.e: Alt+~). 
It should only be applicable for alt+tab when the setting for alt+tab auto-switch on quick release is on.

But is it even needed? commenting it out doesn't seem to circumvent expected functioning.